% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_state}
\alias{assert_state}
\title{Assert that \code{file_path} is a valid topic model state}
\usage{
assert_state(x)
}
\arguments{
\item{x}{Object to check if it is a valid topic_model_state object}
}
\description{
Assert that \code{file_path} is a valid topic model state
}
\details{
The topic model state is a \code{tibble} dataset with at least the variables
\code{doc}, \code{pos}, \code{type} and \code{topic}.
}
\keyword{internal}

