% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_checking.R
\name{mi}
\alias{mi}
\title{Calculate Mutual information between types and documents}
\usage{
mi(state)
}
\arguments{
\item{state}{A topic model state file}
}
\description{
Function to calculate mutual information between types and documents (MI(D,W|k))
for a given topic.

See reference for details.
}
\references{
Mimno, D. and Blei, D. Bayesian Checking for Topic Models
}

