% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_types.R
\name{sample_types_given_topic}
\alias{sample_types_given_topic}
\title{Sample types for a given setup of topic indicators}
\usage{
sample_types_given_topic(state)
}
\arguments{
\item{state}{a topic model state object}
}
\value{
a type vector of length of the state dataset.
}
\description{
The function is sampling new word types based on posterior mode without hyper parameters.

This function is used for modeling comparisons using Bayesian checking.
See reference for details.
}
\references{
Mimno, D. and Blei, D. Bayesian Checking for Topic Models
}

