% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_summaries.R
\name{top_j_types}
\alias{top_j_types}
\title{Return top j terms by weight}
\usage{
top_j_types(wtttm, j)
}
\arguments{
\item{j}{top j types to return (min 1 type)}

\item{wttm}{a weighted tidy type topic matrix}
}

